DatabaseInterface/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
psycopg2==2.9.3
requests==2.27.1
pika==1.2.1
